consumer/requirements.txt
kafka-python==2.0.1
Flask==1.1.2
gunicorn==20.0.4
Flask-SocketIO==4.3.1
